refactor: move uninstall extension action to extension actions
