+++ vicinae/src/actions/root-search/root-search-actions.cpp
@@ -125,23 +125,3 @@ DisableItemAction::DisableItemAction(const QString &id)
     : AbstractAction("Disable item", ImageURL::builtin("trash")), m_id(id) {
   setStyle(AbstractAction::Danger);
 }
-
-void UninstallExtensionAction::execute(ApplicationContext *ctx) {
-  auto alert = new CallbackAlertWidget();
-
-  alert->setTitle("Are you sure?");
-  alert->setMessage(
-      "All this extension data will be permanently lost. If you "
-      "just want the extension to not appear in the root search anymore, consider disabling it instead.");
-  alert->setConfirmText("Uninstall", SemanticColor::Red);
-  alert->setCallback([ctx, id = m_id](bool ok) {
-    if (!ok) return;
-    if (ctx->services->extensionRegistry()->uninstall(id)) {
-      ctx->services->toastService()->setToast("Extension uninstalled");
-    } else {
-      ctx->services->toastService()->setToast("Failed to uninstall extension", ToastPriority::Danger);
-    }
-  });
-
-  ctx->navigation->setDialog(alert);
-}
